racket/draw: avoid callback-based PNG writing
Doc rendering has been occasionally crashing in Pixman/Cairo. It might
have something to do with a non-atomic callback, so let's try avoiding
that, first. (If that doesn't work, the next step is to avoid using
Cairo in a multi-threaded way, for which this is a first step, too.)

original commit: ec6f3fd6106b1285a594c3732d31d3b63e0bc779