racket/draw: clean up object name hacks
The only observable impact of this change is that
`the-color-database` will now print as
`(object:color-database% ...)` instead of
`(object:color-database<%> ...)`. The former makes more
sense anyway, given that you can't have an instance of
an interface. If we really want the latter though,
we can add `object-rename` for this purpose.

original commit: 80ca36e6ab49b3f7061463ad08c0094ca7ca12f4