racket/draw: fix problem with text drawing and scale
Track the transformation of a text-drawing context and reset it
when the current transform changes.

There was already an update on an existing layout for a given
character, but not an update for the context used to create layouts.

Merge to v6.0

original commit: c27c26fb16a9f9b9ef6eb3d9694318a96e51e126