Have record-dc% keep track of font hinting.
Font hinting got added to font% later, and record-dc% was not updated.

This manifested as record-dc%s (such as the ones used for the DrRacket
interactions window) rendering code picts (which turn on hinting)
differently from other drawing backends (bitmaps, slideshow, etc.),
and incorrectly (text widths computed when creating the record-dc%, so
with hinting on, then rendered at a different width, with hinting off).