fix PNG & JPEG callbacks for Racket CS
The reading and writing operations for PNG and JPEG streams involve
C-level callbacks that would like to read and write with arbitrary
ports and sometimes raise an exception. That can work with the
traditional Racket VM, because it can deal with foreign stack frames
in a continuation. Racket CS cannot, so we have to make the callbacks
atomic, add extra jumping for exceptions, and read/write ports
non-incrementally before/after the callback-triggering sequence.