Don’t draw diagonal lines instead of points at high scaling factors
draw-point (and draw-line with a line length of zero) actually draws a
very small line at a 45° angle, which forces Cairo to draw the line cap.
But at large scaling factors, this tiny line stops looking like a point
and starts looking like an actual line, which isn’t good.

This commit adjusts the length of the tiny line based on the current
scaling factor, so it will still look like a point even when scaled up.