Align both brush and pen drawing in draw-ellipse
For some reason, draw-ellipse went out of its way to *not* align
coordinates when drawing with the brush, but that doesn’t make a lot of
sense. Not only is it weird and inconsistent, it could cause problems
where the stroke would actually get drawn inside the boundary of the
ellipse, leaving part of the fill peeking out from “under” the edge.

This just aligns both sets of coordinates, which is more consistent and
fixes the problem.