Allow `make-font` to look up a font in a `font-list%` instance
This allows the dramatically more convenient interface of `make-font` to
be combined with the caching of `font-list%`. Though adding a separate
`find-or-make-font` function might be nominally slightly clearer,
allowing `make-font` to perform both tasks makes it much easier for
functions that create fonts to optionally accept a font list, which is a
useful thing to do.