font%: Add an optional `in-pixels?` argument to `get-size`
This makes it more convenient for code that inspects font sizes to
retrieve them in consistent units.

This commit also fixes `font%` not converting its `size-in-pixels?`
initialization argument to a boolean before retaining it, since that was
almost certainly just an oversight. Also, it adds a missing contract for
`get-feature-settings` that should have been added in
b22958d9fb9793c2ffe3e93ac892336239d6e89a but was overlooked.